Add an experimental DirectoryWatcher that watches for changes to the music folder. After changes are detected, it will trigger a rescan of the music library.
Changed primary key of Track to String. The key will now be musicBrainzId if available or "$artist - $album - $title" otherwise. This should make it possible to re-arrange your library on the file system without affecting playlists.

MusicScanner will now detect when a track is missing its file in the filesystem. In such a case, tracks are now flagged 'missingFile'.

Track .equals and .hashcode methods now compare most of their fields to aid with MusicScanner's ability decide if a track has changes that need to be persisted.

Add visual indicator for tracks with a missing file. Add new favicon.